Create Short Circuit controllers and service
